fix: the DEFAULT made Trajectory Balance unsatisfiable; replace 5 scattered throws with one declarative check
FIRST-PRINCIPLES ROOT CAUSE, not another patched symptom.

I had been fixing demos one at a time -- give this one LEARNABLE_ESTIMATION, give
that one a flow estimator -- without asking why so many needed the same fix. Derived
from the definition instead:

The TB residual is (log Z + sum log P_F - log R - sum log P_B)^2. With Z pinned to 1
it is zero only if P_F(tau) = R(x) P_B(tau|x) for every trajectory. Sum that over all
trajectories reaching x and you get p(x) = R(x), so sum_x p(x) = sum_x R(x). The left
side is 1 because p is a distribution. So TB with a fixed Z is SATISFIABLE ONLY IF
sum_x R(x) = 1.

SIMPLE_ESTIMATION pins Z = 1, and it was the DEFAULT in both TrainingConfig and
create_grid_world_gflownet. So the library's primary objective was provably
unsatisfiable out of the box, and all 26 callers that took the default (12 test files,
14 examples) were silently degraded. The defect was never the 26 callers.

Measured on one demo, changing only this: mean reward went from 8.125 (untrained) ->
8.125 (trained), a 1.0x no-op with the loss RISING 28.021 -> 36.939, to 3.025 ->
15.575, a 5.1x improvement. test_flow_matching_comprehensive.jl showed the same shape
earlier: TB stuck at 21.17 after 200 iterations and WORSE at 1000 (25.46), because
sharpening a policy only grows a residual it can never close.

Default is now LEARNABLE_ESTIMATION in both places. SIMPLE_ESTIMATION remains for the
case it is actually right for -- rewards that sum to 1. Cost: one scalar parameter.

ONE DECLARATIVE CHECK REPLACES FIVE SCATTERED THROWS

Requirements had accumulated as five separate `throw`s buried inside loss branches
(balance.jl:105, balance.jl:128, losses.jl:328, losses.jl:423, training.jl:102), so a
missing component was discovered by crashing on iteration 1 -- or worse, by
train_gflownet catching the exception and recording NaN for every iteration, which is
how several demos "completed" while learning nothing.

validate_training_config now checks one declarative table up front and reports EVERY
missing piece at once with the mathematical reason. It could not have been called
before: it read `model.dag.states`, and GFlowNetModel has no `dag` field, so it threw
on any model. That is why nothing called it. It was also wrong where it did work, and
get_objective_requirements disagreed with it -- claiming SUB_TRAJECTORY_BALANCE needs
only a forward policy when it needs four things. That function is now DERIVED from the
same table, so the two cannot drift again.

TWO SEVERITIES, split on measured evidence rather than taste:
  ERROR   SubTB without a learnable Z -- collapses onto ONE terminal state with
          probability 1.000 on every seed tried, TV 0.9474. Degenerate.
  WARNING TB or TLM without a learnable Z -- unsatisfiable but still trains, still
          moves the policy, and is outright correct when rewards sum to 1. Refusing
          would break legitimate code-path tests for no mathematical gain.

Corrected two of my own over-claims while getting the suite green: DB's flow estimator
is OPTIONAL, since losses.jl falls back to the recursive flow, and I had clobbered the
FLOW_MATCHING branch entirely in an edit.

ELTYPE REGRESSION THE DEFAULT CHANGE EXPOSED

log_Z was inserted as a Float64 into a Float32 ComponentArray, promoting EVERY layer
to Float64 -- double memory and different numerics. Latent until LEARNABLE became the
default, then surfaced as compute_forward_logits returning Vector{Float64} and failing
test_core_functions.jl:131. Now derived from the network parameter eltype, so a Float32
model stays Float32 and a Float64 model stays Float64. Also had to convert at
validation.jl:316, where ps.log_Z is passed to a field typed Union{Nothing,Float64}.

Suite: 1859 passed, 0 failed, 0 errored. Broken went 48 -> 28 and passing 1257 -> 1859,
i.e. 602 assertions that previously did not execute now do.

+++ src/utils/validation.jl
@@ -307,7 +307,13 @@ function validate_z_gradients(model::GFlowNetModel, trajectories::Vector{Traject
             model.forward_policy,
             model.backward_policy,
             model.flow_estimator,
-            haskey(ps, :log_Z) ? ps.log_Z : nothing,
+            # Float64(...) is REQUIRED: the struct field is Union{Nothing,Float64}
+            # (types.jl:160) while ps.log_Z follows the network parameter eltype,
+            # which is Float32. Passing it straight through threw
+            # MethodError: no method matching GFlowNetModel(..., ::Float32, ...).
+            # The field is a mirror of the parameter, so widening it is unnecessary;
+            # converting here keeps the struct concretely typed.
+            haskey(ps, :log_Z) ? Float64(ps.log_Z) : nothing,
             ps,
             model.optimizer,
             model.states
